Use server-side state for charms by default
Storing data in the SQLite DB is a legacy method of storage which also
requires persistent volumes for operators themselves. Upon pod
recreation the unit-local state is lost.

It would be good to either use the newer method which does not rely on
persistent volumes or provide a recommendation in metadata.yaml on how
to configure this properly since this is a template. This change ops for
the former.